pyproject.toml
[project]
name = "repl-nix-q-connect-webapp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.3",
    "plotly>=5.24.1",
    "qiskit-aer>=0.15.1",
    "qiskit>=1.2.4",
    "streamlit>=1.40.1",
    "components>=0.0.1a0",
    "streamlit-authenticator>=0.4.1",
    "pyyaml>=6.0.2",
    "bcrypt>=4.2.1",
    "reportlab>=4.2.5",
]
